Intepret point statements
